criada classe servletUtil

+++ page-jsp/src/main/java/DAO/DAOUsuarioRepository.java
@@ -16,71 +16,97 @@ public class DAOUsuarioRepository {
 	public DAOUsuarioRepository() {
 
 		connection = SingleConnectionBanco.getConnection();
+		
 
 	}
 
-	public ModelLogin gravarUser(ModelLogin objeto) throws Exception {
-		
-		if(objeto.isNovo()) {/*Grava um novo*/
-		
+	public ModelLogin gravarUser(ModelLogin objeto, Long userLogado) throws Exception {
 
-		String sql = " INSERT INTO model_login (login, senha, nome, email)  VALUES (?, ?, ?, ?)";
+		if (objeto.isNovo()) {/* Grava um novo */
 
-		PreparedStatement preparedSql = connection.prepareStatement(sql);
-		preparedSql.setString(1, objeto.getLogin());
-		preparedSql.setString(2, objeto.getSenha());
-		preparedSql.setString(3, objeto.getNome());
-		preparedSql.setString(4, objeto.getEmail());
+			String sql = " INSERT INTO model_login (login, senha, nome, email, usuario_id)  VALUES (?, ?, ?, ?, ?)";
 
-		preparedSql.execute();
-		connection.commit();
+			PreparedStatement preparedSql = connection.prepareStatement(sql);
+			preparedSql.setString(1, objeto.getLogin());
+			preparedSql.setString(2, objeto.getSenha());
+			preparedSql.setString(3, objeto.getNome());
+			preparedSql.setString(4, objeto.getEmail());
+			preparedSql.setLong(5, userLogado);
 
-		}else {
-			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id = "+objeto.getId()+" ;";
+			preparedSql.execute();
+			connection.commit();
+
+		} else {
+			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id = " + objeto.getId() + " ;";
 			PreparedStatement prepareSql = connection.prepareStatement(sql);
-			
+
 			prepareSql.setString(1, objeto.getLogin());
 			prepareSql.setString(2, objeto.getSenha());
 			prepareSql.setString(3, objeto.getNome());
 			prepareSql.setString(4, objeto.getEmail());
-			
+		
 			prepareSql.executeUpdate();
 			connection.commit();
-			
+
 		}
-		
-		return this.consultaUsuario(objeto.getLogin());
- 
+
+		return this.consultaUsuario(objeto.getLogin(), userLogado);
+
 	}
 	
 	
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception{
-		
-		List<ModelLogin> retorno = new ArrayList<>();
-		
-		 String sql = "select * from model_login where upper (nome) like upper(?)";
-		 PreparedStatement statement = connection.prepareStatement(sql);
-		 statement.setString(1, "%"+nome+"%");
-		 
-		 ResultSet resultado = statement.executeQuery();
-		 
-		 while(resultado.next()) {/*precorrer as linhas de resultado SQL*/
-			 ModelLogin modelLogin = new ModelLogin();
-			 modelLogin.setId(resultado.getLong("id"));
-			 modelLogin.setNome(resultado.getString("nome"));
-			 modelLogin.setEmail(resultado.getString("email"));
-			 modelLogin.setLogin(resultado.getString("login"));
-			 
-			 retorno.add(modelLogin);
-		 }
-		 
-		
-		
+
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception {
+
+		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
+
+		String sql = "select * from model_login where useradmin is false and usuario_id = " + userLogado;
+		PreparedStatement statement = connection.prepareStatement(sql);
+
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) {/* precorrer as linhas de resultado SQL */
+
+			ModelLogin modelLogin = new ModelLogin();
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+
+			retorno.add(modelLogin);
+		}
+
 		return retorno;
-		
+
 	}
 
-	public ModelLogin consultaUsuario(String login) throws Exception {
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception {
+
+		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
+
+		String sql = "select * from model_login where upper (nome) like upper(?) and useradmin is false and username is false usuario_id = ?";
+		PreparedStatement statement = connection.prepareStatement(sql);
+		statement.setString(1, "%" + nome + "%");
+		statement.setLong(2, userLogado);
+
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) {/* precorrer as linhas de resultado SQL */
+			ModelLogin modelLogin = new ModelLogin();
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+
+			retorno.add(modelLogin);
+		}
+
+		return retorno;
+
+	}
+	
+	public ModelLogin consultaUsuarioLogado(String login) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
@@ -97,6 +123,83 @@ public class DAOUsuarioRepository {
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setSenha(resultado.getString("senha"));
+			
+
+		}
+
+		return modelLogin;
+
+	}
+	
+	public ModelLogin consultaUsuario(String login) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "select * from model_login where upper(login) = upper('" + login + "') and useradmin is false ";
+
+		PreparedStatement statement = connection.prepareStatement(sql);
+
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) {/* se tem resultado */
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			
+
+		}
+
+		return modelLogin;
+
+	}
+
+	public ModelLogin consultaUsuario(String login, Long userLogado) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "select * from model_login where upper(login) = upper('" + login + "') and useradmin is false and usuario_id = " + userLogado;
+
+		PreparedStatement statement = connection.prepareStatement(sql);
+
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) {/* se tem resultado */
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			
+
+		}
+
+		return modelLogin;
+
+	}
+
+	public ModelLogin consultaUsuarioID(String id, Long userLogado) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "select * from model_login where id = ? and useradmin is false and usuario_id = ?";
+
+		PreparedStatement statement = connection.prepareStatement(sql);
+		statement.setLong(1, Long.parseLong(id));
+		statement.setLong(2, userLogado);
+
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) {/* Se Tem Resultado */
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setSenha(resultado.getString("senha"));
 
 		}
 
@@ -105,10 +208,6 @@ public class DAOUsuarioRepository {
 	}
 
 	public boolean validarLogin(String login) throws Exception {
-		
-		
-			
-		
 
 		String sql = "select count(1) > 0 as existe from model_login where upper(login) = upper('" + login + "')";
 
@@ -120,21 +219,19 @@ public class DAOUsuarioRepository {
 		return resultado.getBoolean("existe");
 
 	}
-	
-	public  void deletarUser(String idUser) throws Exception {
-		
-		String sql = " DELETE FROM model_login WHERE id = ? ;";
-		
+
+	public void deletarUser(String idUser) throws Exception {
+
+		String sql = " DELETE FROM model_login WHERE id = ? and useradmin is false;";
+
 		PreparedStatement prepareSql = connection.prepareStatement(sql);
-		
+
 		prepareSql.setLong(1, Long.parseLong(idUser));
-		
+
 		prepareSql.executeUpdate();
-		
+
 		connection.commit();
-		
-		
+
 	}
-	
-	
+
 }
